% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{adjust_Gram}
\alias{adjust_Gram}
\title{Function to get Gram matrix after adjusting of sigma parameter of a kernel}
\usage{
adjust_Gram(kernel, sigma = (2^(1:20)) * 0.001, x, y)
}
\arguments{
\item{kernel}{Function of kernel with parameter sigma, class from kernel lab package}

\item{sigma}{numeric vector of possible values of the sigma parameter for a kernel function,
by default \code{sigma = ( 2**(1:20) ) * 1E-3}}

\item{x}{Matrix of stat.sim}

\item{y}{Matrix of stat.obs}
}
\value{
\code{adjust_Gram} function returns Gram matrix after adjusting of sigma
}
\description{
Function to get Gram matrix after adjusting of sigma parameter of a kernel
}
\examples{
NULL 
}
