% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{analyze_experiments}
\alias{analyze_experiments}
\title{Function to get statistics from the results of experiments and find the best methods for each simulation case}
\usage{
analyze_experiments(DF, file_to_save = "../gplot.pdf")
}
\arguments{
\item{DF}{Data frame with results of experiments, output of \code{experiment_models()} function.}
}
\value{
List of statistical data with analysis of results: \cr
\itemize{
\item ; \cr
\item ; \cr
\item ; \cr
\item ; \cr
\item ; \cr
\item ; \cr
\item .
}
}
\description{
Function to get statistics from the results of experiments and find the best methods for each simulation case
}
\examples{
NULL 
}
