% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_range}
\alias{apply_range}
\title{Function to restrict values of the data according with the range for each dimension}
\usage{
apply_range(diapason, input.data)
}
\arguments{
\item{diapason}{Vector of min and max values or data frame with two rows (min and max) for each dimension of input data}

\item{input.data}{Data frame of input where values will be corrected}
}
\value{
The same data frame with corrected values according to the diapason
}
\description{
Function to restrict values of the data according with the range for each dimension
}
\examples{
MaxWiK::MaxWiK_templates(dir = './') # See the templates and vignettes for usage
}
