% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric_format}
\alias{check_numeric_format}
\title{Function to check DATA.FRAME}
\usage{
check_numeric_format(l)
}
\arguments{
\item{l}{DATA.FRAME that should have data of numeric type}
}
\value{
TRUE if data.frame has ONLY numeric data and FALSE vice verse
}
\description{
Check that DATA.FRAME has numeric format for ALL the columns and it has NO 'NA' values
}
\examples{
\dontrun{
check_numeric_format( data.frame( A= c(9,0), B = c(4,6)) )  # TRUE
check_numeric_format( data.frame( A= c(9,0), B = c(4,NA)) )  # Error due to NA value
check_numeric_format( data.frame( A= c(9,'0'), B = c(4,6)) ) # Error due to character in the data
}
}
