% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Check the installation of packages and attach them with corresponding functions}
\usage{
check_packages(pkgs = NULL)
}
\arguments{
\item{pkgs}{List of package names with related function names,
by default (or when pkgs = NULL) the list of packages are described in Namespace file of the package or
'R/MaxWiK-package.R' file}
}
\value{
if the packages are installed then it returns NULL else it returns error message
}
\description{
Check the installation of packages and attach them with corresponding functions
}
\examples{
NULL
}
\keyword{internal}
