% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_pipelines}
\alias{copy_pipelines}
\title{Function to copy the pipelines from extdata folder in the library to /Pipelines/ folder in the working directory}
\usage{
copy_pipelines(dir = "./")
}
\arguments{
\item{dir}{Folder to where files should be save, by default dir = './'}
}
\value{
List of logic numbers for each copied file, TRUE - success, FALSE - not success
}
\description{
Function to copy the pipelines from extdata folder in the library to /Pipelines/ folder in the working directory
}
\examples{
copy_pipelines( dir = 'Input' )
}
