% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_templates}
\alias{copy_templates}
\title{Function to copy the templates from extdata folder in the library to /Templates/ folder in the working directory}
\usage{
copy_templates(dir = "./")
}
\arguments{
\item{dir}{Folder to where files should be save, by default dir = './'}
}
\value{
List of logic numbers for each copied file, TRUE - success, FALSE - not success
}
\description{
Function to copy the templates from extdata folder in the library to /Templates/ folder in the working directory
}
\examples{
copy_templates( dir = './' )
}
