% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiment_models}
\alias{experiment_models}
\title{Function to prepare toy experiments}
\usage{
experiment_models(
  file_name = "output.txt",
  models = c("Gaussian", "Linear"),
  dimensions = (1:20) * 2,
  stochastic_terms = c(0, 0.1, 0.3, 0.7, 1, 1.5),
  rng = c(0, 10),
  restrict_points_number = 300
)
}
\arguments{
\item{file_name}{Name of file to output results}

\item{models}{Names of models for simulation, by default
\code{models = c( 'Gaussian', 'Linear' )}}

\item{dimensions}{Dimensions of models, by default
\code{ dimensions = (1:20)*2 }}

\item{stochastic_terms}{Stochastic terms for each model, by default
\code{ stochastic_terms  =  c( 0, 0.1, 0.3, 0.7, 1 ) }}

\item{rng}{Range for each variable, by default
\code{rng  =  c( 0,10 ) }}
}
\value{
\code{experiment_models()} returns list of results of experiments
}
\description{
Function to prepare toy experiments
}
\examples{
NULL
}
