% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{Gaussian_kernel}
\alias{Gaussian_kernel}
\alias{Laplacian_kernel}
\title{Kernel function for two vectors x1 and x2}
\usage{
Gaussian_kernel(x1, x2, sgm = 1)

Laplacian_kernel(x1, x2, b = 1)
}
\arguments{
\item{x1}{Numeric vector x1}

\item{x2}{Numeric vector x2}

\item{sgm}{Sigma parameter in Gaussian kernel}

\item{b}{The width of Laplacian kernel}
}
\value{
Value of Gaussian kernel for two vectors and given sigma

Value of Laplacian kernel for two vectors and given width
}
\description{
Kernel function for two vectors x1 and x2
}
\section{Functions}{
\itemize{
\item \code{Laplacian_kernel}: Value of Laplacian kernel for two vectors and given width
}}

\examples{
Gaussian_kernel(x1 = c(3,4), x2 = c(3.4,5), sgm = 1 )
Gaussian_kernel(x1 = c(3,4), x2 = c(3.4,5), sgm = 5 )
Gaussian_kernel(x1 = c(3,4), x2 = c(4,5), sgm = 1 )
Laplacian_kernel(x1 = c(3,4), x2 = c(4,5), b = 1) 
Laplacian_kernel(x1 = c(3,4), x2 = c(4,5), b = 5) 
Laplacian_kernel(x1 = c(3,4), x2 = c(3,5), b = 1) 
}
