% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{Gaussian_model}
\alias{Gaussian_model}
\alias{linear_model}
\title{The model of simulations that is based on Gaussian functions for each dimension}
\usage{
Gaussian_model(
  d = 1,
  x0 = 3,
  probability = TRUE,
  n = 1000,
  r = range(0, 10),
  noise = 0
)

linear_model(
  d = 1,
  x0 = 3,
  probability = TRUE,
  noise = 0.2,
  n = 1000,
  r = range(0, 10)
)
}
\arguments{
\item{d}{Dimension of the parameter and model space}

\item{x0}{Numeric vector with length of dimensionality of data frame,
that contents the truth value of parameter. Each number in the vector should be within the range r}

\item{probability}{Logical, if TRUE then apply uneven distribution for parameters generation}

\item{n}{Integer number of points in data frames}

\item{r}{Range \code{r = c(min, max)}, by default  \code{r = range(0,10)}}

\item{noise}{Noise factor, implemented as coefficient in \code{f(x) = ... + noise * runif(1)}}
}
\value{
The function \code{Gaussian_model()} returns list of two objects: \cr
\itemize{
\item stat.sim - data frame of simulations statistics,
\item par.sim - data frame of parameters,
\item stat.obs - data frame of an observation point.
}

The function \code{linear_model()} returns list of two objects: \cr
\itemize{
\item stat.sim - data frame of simulations statistics,
\item par.sim - data frame of parameters,
\item stat.obs - data frame of an observation point.
}
}
\description{
The function \code{Gaussian_model()} allows to generate
parameters and statistics of simulations that are based on Gaussian function for each dimension: \cr
\code{f( x ) = { exp( - (x1-x01)^2 / 2 ), ..., exp( - (xn-x0n)^2 / 2 )  } = { y1, y2, ..., yn } } is a vector of output data

The function \code{linear_model()} allows to generate
parameters and statistics of simulations that are based on linear function for each dimension: \cr
\code{f( x ) = { 1  +  ( x1 - x01 ) / x01 + noise_1, ..., 1  +  ( xn - x0n ) / x0n + noise_n  } = { y1, y2, ..., yn } } is a vector of output data
}
\section{Functions}{
\itemize{
\item \code{linear_model()}: The model of simulations that is based on linear functions for each dimension

}}
\examples{
NULL
NULL
}
