% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iKernelABC.R
\name{get.iKernelABC}
\alias{get.iKernelABC}
\title{Function to get Approximate Bayesian Computation based on Maxima Weighted Isolation Kernel mapping}
\usage{
get.iKernelABC(
  psi = 40,
  t = 350,
  param,
  stat.sim,
  stat.obs,
  talkative = FALSE,
  check_pos_def = TRUE
)
}
\arguments{
\item{psi}{Integer number. Size of each Voronoi diagram or number of areas/points in the Voronoi diagrams}

\item{t}{Integer number of trees in the Isolation Forest}

\item{param}{or \code{par.sim} - data frame of parameters of the model}

\item{stat.sim}{Summary statistics of the simulations (model output)}

\item{stat.obs}{Summary statistics of the observation point}

\item{talkative}{Logical parameter to print or do not print messages}

\item{check_pos_def}{Logical parameter to check the Gram matrix is positive definite or do not check}
}
\value{
The function \code{get.iKernelABC()} returns the list of :
\itemize{
\item kernel_mean_embedding is a maxima weighted kernel mean embedding (mapping) related to the observation point;
\item parameters_Matrix_Voronoi is a matrix of information about Voronoi trees (rows - trees, columns - Voronoi points/areas IDs) for parameters data set;
\item parameters_Matrix_iKernel is a matrix of of all points of PARAMETERS in a Hilbert space (rows - points, columns - isolation trees);
\item Hilbert_weights is a weights in Hilbert space to get maxima weighted kernel mean embedding for parameters_Matrix_iKernel;
\item Matrix_iKernel is a matrix of all points of simulations in a Hilbert space (rows - points, columns - isolation trees);
\item iFeature_point is a feature embedding mapping for the OBSERVATION point;
\item similarity is a vector of similarities between the simulation points and observation point;
\item Matrix_Voronoi is a matrix of information about Voronoi trees (rows - trees, columns - Voronoi points/areas IDs);
\item t is a number of trees in the Isolation Forest;
\item psi is a number of areas/points in the Voronoi diagrams
}
}
\description{
The function \code{get.iKernelABC()} is used to get Approximate Bayesian Computation
based on Maxima Weighted Isolation Kernel mapping.
On given data frame of parameters, statistics of the simulations and an observation,
using the internal parameters psi and t,
the function \code{get.iKernelABC()} returns the estimation of a parameter corresponding to
Maxima weighted Isolation Kernel ABC method.
}
\examples{
NULL
}
