% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{get_dataset_of_Gaussian_model}
\alias{get_dataset_of_Gaussian_model}
\alias{Gauss_function}
\title{Function to get dataset of simulations that is based on Gaussian functions for each dimension}
\usage{
get_dataset_of_Gaussian_model(
  d = 1,
  x0 = 3,
  r = range(0, 10),
  noise = 0,
  A = 1,
  sigma = 1,
  probability = TRUE,
  n = 1000
)

Gauss_function(d = 1, x0 = 3, r = range(0, 10), noise = 0, A = 1, sigma = 1, x)
}
\arguments{
\item{d}{Dimension of the parameter and model space}

\item{x0}{Numeric vector with length of dimensionality of data frame,
that contents the truth value of parameter. Each number in the vector should be within the range r}

\item{r}{Range \code{r = c(min, max)}, by default  \code{r = range(0,10)} for all the dimensions}

\item{noise}{Number corresponding to noise level in the dataset. 0 means the absent of noise,
1 means noise level is the same as amplitude, more than 1 means that a noise is greater than 'signal'.}

\item{A}{Vector of exponent multiplication factors}

\item{sigma}{Vector of sigmas in Gaussian function. Length of vector should be equal d.}

\item{probability}{Logical, if TRUE then apply uneven distribution for parameters generation}

\item{n}{Integer number of points in data frame}
}
\value{
The function \code{get_dataset_of_Gaussian_model()} returns list of three objects: \cr
\itemize{
\item stat.sim - data frame of simulations statistics,
\item par.sim - data frame of parameters,
\item stat.obs - data frame of an observation point.
}

\code{Gauss_function()} returns output of Gaussian function for d dimensions
}
\description{
The function \code{get_dataset_of_Gaussian_model()} allows to generate
parameters and statistics of simulations that are based on Gaussian function for each dimension: \cr
\code{f( x ) = { exp( - (x1-x01)^2 / 2 ), ..., exp( - (xn-x0n)^2 / 2 )  } = { y1, y2, ..., yn } } is a vector of output data

The function \code{Gauss_function()} is used as
a model based on Gaussian function for multidimensional case.
}
\section{Functions}{
\itemize{
\item \code{Gauss_function()}: Function to get output of Gaussian function for d dimension case

}}
\examples{
NULL
NULL
}
