% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{get_dataset_of_Linear_model}
\alias{get_dataset_of_Linear_model}
\alias{Linear_function}
\title{Function to get dataset of simulations that is based on Linear functions for each dimension}
\usage{
get_dataset_of_Linear_model(
  d = 1,
  x0 = 3,
  r = range(0, 10),
  noise = 0,
  A = 1,
  probability = TRUE,
  n = 1000
)

Linear_function(d = 1, x0 = 3, r = range(0, 10), noise = 0, A = 1, x)
}
\arguments{
\item{d}{Dimension of the parameter and model space}

\item{x0}{Numeric vector with non-zero values. The vector's length should be equal d.
\code{x0} contents the truth value of parameter. Each number in the vector should be within the range r}

\item{r}{Range \code{r = c(min, max)}, by default  \code{r = range(0,10)} for all the dimensions}

\item{A}{Vector of multiplication factors.}

\item{probability}{Logical, if TRUE then apply uneven distribution for parameters generation}

\item{n}{Integer number of points in data frame}
}
\value{
The function \code{get_dataset_of_Linear_model()} returns list of three objects: \cr
\itemize{
\item stat.sim - data frame of simulations statistics,
\item par.sim - data frame of parameters,
\item stat.obs - data frame of an observation point.
}

\code{Linear_function()} returns output of Linear function for d dimensions
}
\description{
The function \code{get_dataset_of_Linear_model()} allows to generate
parameters and statistics of simulations that are based on Linear function for each dimension: \cr
\code{f( x ) = { A1 * x1 / x01, ..., An * xn / x0n  } = { y1, y2, ..., yn } } is a vector of output data

The function \code{Linear_function()} is used as
a model based on Linear function for multidimensional case.
}
\section{Functions}{
\itemize{
\item \code{Linear_function()}: Function to get output of Linear function for d dimension case

}}
\examples{
NULL
NULL
}
