% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{get_kernel_mean_embedding}
\alias{get_kernel_mean_embedding}
\title{The function to calculate Maxima weighted kernel mean mapping for Isolation Kernel in RKHS related to parameters space}
\usage{
get_kernel_mean_embedding(parameters_Matrix_iKernel, Hilbert_weights)
}
\arguments{
\item{parameters_Matrix_iKernel}{Matrix of all the points represented in RKHS related to parameters space}

\item{Hilbert_weights}{Maximal weights in RKHS to get related part of kernel mean embedding from parameters_Matrix_iKernel}
}
\value{
Maxima weighted kernel mean mapping in the form of integer vector with length t (number of trees).
Each element of the vector is index of Voronoi cell with maximal weight in the Voronoi diagram
}
\description{
The function to calculate Maxima weighted kernel mean mapping for Isolation Kernel in RKHS related to parameters space
}
\examples{
NULL
}
