% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script-MSE.R
\name{Get_MAP}
\alias{Get_MAP}
\title{The function to get Maximum A Posteriori from numeric data frame}
\usage{
Get_MAP(DF)
}
\arguments{
\item{DF}{Data frame of the integer or numeric numbers or characters}
}
\value{
The function \code{Get_MAP} returns the MAP for each dimension in data frame DF of vector
}
\description{
The function \code{Get_MAP} returns the Maximum A Posteriori (MAP) of data frame.
}
\examples{
NULL 
}
