% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{Get_Mode}
\alias{Get_Mode}
\title{The function to get mode of the input vector}
\usage{
Get_Mode(v)
}
\arguments{
\item{v}{Input vector of the integer, numeric numbers or characters}
}
\value{
The function \code{Get_Mode} returns the mode of vector
}
\description{
The function \code{Get_Mode} returns the mode of vector. \cr
The mode is the value that appears most often in a dataset.
}
\examples{
NULL 
}
