% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{get_Spider_MAP}
\alias{get_Spider_MAP}
\title{Function to get MAP of SpiderWeb algorithm based on different psi / t hyperparameters}
\usage{
get_Spider_MAP(
  stat.sim,
  par.sim,
  stat.obs,
  restrict_points_number = 300,
  cores = 4,
  n_best = 8
)
}
\arguments{
\item{stat.sim}{Summary statistics of the simulations (model output)}

\item{par.sim}{\itemize{
\item data frame of parameters of the model
}}

\item{stat.obs}{Summary statistics of the observation point}

\item{restrict_points_number}{Maximal number of points in the data sets to get MAP}

\item{cores}{Number of cores for parallel calculation}

\item{n_best}{Number of best psi_t pairs adjusted for MaxWiK algorithm}
}
\value{
Maximum A Posteriori of meta-sampling distribution of parameters
}
\description{
Function to get MAP of SpiderWeb algorithm based on different psi / t hyperparameters
}
\examples{
NULL
# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
sim = simulation_example_many_psi_t( verbose = FALSE , to_plot = FALSE )
}
