% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{GET_SUBSET}
\alias{GET_SUBSET}
\title{The function to get subset with size psi for Voronoi diagram}
\usage{
GET_SUBSET(data_set, pnts)
}
\arguments{
\item{data_set}{Data.frame of Voronoi diagram}

\item{pnts}{Integer vector of indexes of columns of the data_set}
}
\value{
Subset of data_set with columns pnts
}
\description{
The function to get subset with size psi for Voronoi diagram
}
\examples{
NULL

}
