% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{get_subset_of_feature_map}
\alias{get_subset_of_feature_map}
\title{The function to get subset of points based on feature mapping}
\usage{
get_subset_of_feature_map(dtst, Matrix_Voronoi, iFeature_point)
}
\arguments{
\item{dtst}{Dataset of all the original points}

\item{Matrix_Voronoi}{Matrix of Voronoi diagrams based on the Isolation Kernel algorithm}

\item{iFeature_point}{Feature mapping in RKHS for a point,
that can be gotten via \code{add_new_point_iKernel()} function}
}
\value{
The subset of dtst that has points
extracted with feature mapping of an observation point (iFeature_point)
}
\description{
The function to get subset of points based on feature mapping
}
\examples{
NULL
}
