% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{get_voronoi_feature_PART_dataset}
\alias{get_voronoi_feature_PART_dataset}
\title{The function to get feature representation in RKHS based on Voronoi diagram for PART of dataset}
\usage{
get_voronoi_feature_PART_dataset(
  data,
  talkative = FALSE,
  start_row,
  Matrix_Voronoi
)
}
\arguments{
\item{data}{Data.frame of new points}

\item{talkative}{Logical parameter to print or do not print messages}

\item{start_row}{Row number from which a new data should be added}

\item{Matrix_Voronoi}{Matrix of Voronoi diagrams based on the PREVIOUS dataset}
}
\value{
List of three matrices: Matrix_Voronoi, Matrix_iKernel and dissim
}
\description{
\code{get_voronoi_feature_PART_dataset()} function returns
the feature (mapping) representation in RKHS based on Voronoi diagram for NEW PART of dataset.
The \code{Matrix_Voronoi} is based on the PREVIOUS dataset.
The NEW PART of dataset will appear at the end of PREVIOUS dataset
}
\examples{
NULL
}
\keyword{internal}
