% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{K2_ABC}
\alias{K2_ABC}
\title{Function to get parameter estimation and weights using K2-ABC method}
\usage{
K2_ABC(G, epsilon = 0.1, par.sim)
}
\arguments{
\item{G}{Gram matrix between sim.stat and obs.stat data sets/matrices that can be obtained by \cr
\code{ krnl = rbfdot(sigma = 1) } \cr
\code{G = kernelMatrix( kernel = krnl, x = as.matrix(sim.stat), y = as.matrix(obs.stat) ) }}

\item{epsilon}{adjust parameter in the weight estimation: \cr
\code{w_ij = exp(- ( 1 - k_(y_i, y_obs )/epsilon )}, where \code{y_obs } is observation value
and \code{y_i} is a i-th point from sim.stat, by default epsilon = 0.1}

\item{par.sim}{dataset/matrix of parameters for simulation}
}
\value{
\code{K2_ABC()} returns the list of: \cr
\enumerate{
\item weights for \code{par.sim} related to observation point based on Gram matrix \cr
\item parameter estimation par.est
}
}
\description{
\code{K2_ABC()} function allows to get parameter
estimation and weights using K2-ABC method described in
the paper
Mijung Park, Wittawat Jitkrittum, Dino Sejdinovic,
Proceedings of the 19th International Conference on Artificial Intelligence and Statistics,
\href{https://proceedings.mlr.press/v51/park16.html}{PMLR 51:398-407, 2016}.
}
\examples{
NULL
}
