% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{MaxWiK.env}
\alias{MaxWiK.env}
\title{Environment of the package 'MaxWiK' to store all the common objects of a simulation}
\format{
An object of class \code{environment} of length 0.
}
\usage{
MaxWiK.env
}
\description{
\code{MaxWiK.env} is environment of the package 'MaxWiK'
where all the common objects of a simulation are stored and used
}
\keyword{datasets}
