% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My.Plots.R
\name{MaxWiK.ggplot.density}
\alias{MaxWiK.ggplot.density}
\title{Density plot}
\usage{
MaxWiK.ggplot.density(
  title = "",
  datafr1,
  datafr2,
  var.df,
  obs.true = NULL,
  best.sim = NULL,
  clrs = c("#a9b322", "#f9b3a2", "red", "blue"),
  alpha = c(0.1, 0.4),
  lw = c(0.7, 0.7),
  lt = c("dashed", "dotted")
)
}
\arguments{
\item{title}{Title of the plot}

\item{datafr1}{data frame 1}

\item{datafr2}{data frame 2}

\item{var.df}{Variables to show}

\item{obs.true}{True observation if so, NULL by default}

\item{best.sim}{The best point from a simulation if so, NULL by default}

\item{clrs}{Colors to plot, by default it is c( "#a9b322", "#f9b3a2", 'red', 'blue' )}

\item{alpha}{Transparency values for density plots}

\item{lw}{Line widths}

\item{lt}{Line types}
}
\value{
Make and return the ggplot object of the densities of the data frames
}
\description{
Density plot
}
\examples{
MaxWiK::MaxWiK_templates(dir = tempdir()) # See the templates and vignettes for usage. 
# Function 'MaxWiK.ggplot.density()' is used in the MaxWiK.ABC.R and 
# MaxWiK.Predictor.R templates.
}
