% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{Mean_iKernel_parameters}
\alias{Mean_iKernel_parameters}
\title{The function returns the weighted mean of the parameter based on Isolation Kernel ABC method}
\usage{
Mean_iKernel_parameters(param, sm)
}
\arguments{
\item{param}{Data frame of parameters}

\item{sm}{Numeric vector of weights gotten from \code{iKernelABC()} function}
}
\value{
The function \code{Mean_iKernel_parameters()} returns the weighted mean of the parameter
}
\description{
The function \code{Mean_iKernel_parameters()} returns the weighted mean of the parameter
that was calculated with \code{KernelABC()} function that represents Isolation Kernel ABC method
}
\examples{
NULL 
}
