% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{model}
\alias{model}
\title{Model definition for sampler to calculate one simulation}
\usage{
model(
  name = c("Gaussian", "Linear")[1],
  parameter,
  x0,
  stat.obs,
  noise = 0,
  sigma = 1,
  A = 1
)
}
\arguments{
\item{name}{Name of a model, can be either \code{ 'Gaussian' or 'Linear' }}

\item{parameter}{Value of a parameter}

\item{x0}{True value of parameter}

\item{stat.obs}{Data frame of statistics of observation point}

\item{noise}{Value of stochastic term}
}
\value{
\code{model()} returns data frame with a result of a simulation based on the model
}
\description{
Model definition for sampler to calculate one simulation
}
\examples{
NULL
}
