% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{norm_vec}
\alias{norm_vec}
\alias{norm_vec_sq}
\title{The norm function for vector}
\usage{
norm_vec(x)

norm_vec_sq(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The squared root of sum of squared elements of the vector x or Euclid length of the vector x

The squared Euclid norm or the sum of squared elements of the vector x
}
\description{
The norm function for vector
}
\section{Functions}{
\itemize{
\item \code{norm_vec_sq()}: The squared norm or the sum of squared elements of the vector x

}}
\examples{
NULL
NULL
}
\keyword{internal}
