% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plots.R
\name{plot_sudoku_2D}
\alias{plot_sudoku_2D}
\alias{plot_web_2D}
\title{Function to plot all the results of the function \code{sudoku()}}
\usage{
plot_sudoku_2D(
  stat.sim,
  par.sim,
  par.truth,
  iKernelABC,
  rslt,
  ind_X,
  ind_Y,
  names = c("Parameter_1", "Parameter_2"),
  xlim,
  ylim,
  show_tracer = TRUE,
  show_obs = TRUE,
  show_appropriate = TRUE,
  show_best = TRUE,
  show_u_point = TRUE,
  show_legend = FALSE
)

plot_web_2D(
  stat.sim,
  par.sim,
  par.truth,
  iKernelABC,
  web,
  ind_X,
  ind_Y,
  names = c("P1", "P2"),
  xlim,
  ylim,
  show_tracer = TRUE,
  show_obs = TRUE,
  show_network = TRUE,
  show_best = TRUE,
  show_u_point = TRUE,
  show_legend = FALSE
)
}
\arguments{
\item{stat.sim}{Summary statistics of the simulations (model output)}

\item{par.sim}{Data frame of parameters of the model}

\item{par.truth}{Truth value of the parameter corresponding to observation point (if known)}

\item{iKernelABC}{Result of calculations based on Isolation Kernel ABC
that can be gotten by the function \code{iKernelABC()}}

\item{rslt}{Results of function \code{sudoku()}}

\item{ind_X}{Column index of the par.sim data frame to plot as X axes}

\item{ind_Y}{Column index of the par.sim data frame to plot as Y axes}

\item{names}{Vector of axes names, by default \code{names = c( 'Parameter_1', 'Parameter_2' )}}

\item{xlim}{Numeric vector of the range for X axes}

\item{ylim}{Numeric vector of the range for Y axes}

\item{show_tracer}{Logical to show or do not show all the tracer points}

\item{show_obs}{Logical to show or do not show the stat.obs or parameter of observation point (if known)}

\item{show_appropriate}{Logical to show or do not show all the points of
the object \code{rslt$surroundings_best_points} from results of \code{sudoku()} function}

\item{show_best}{Logical to show or do not show the best point of the results of \code{sudoku()} function}

\item{show_u_point}{Logical to show or do not show the estimated point of parameter for an observation}

\item{show_legend}{Logical to show or do not show the legend}

\item{web}{Results of the function \code{spiderweb()} to draw on the plot}

\item{show_top}{Logical to show or do not show the top points of the results of \code{sudoku()} function}
}
\value{
Plot all the results of \code{sudoku()} function

The function \code{plot_web_2D()} draws all the results of the function \code{spiderweb()}
}
\description{
The function \code{plot_sudoku_2D} draws all the results of the function \code{sudoku()}

The function \code{plot_web_2D()} draws all the results of the function \code{spiderweb()}
}
\section{Functions}{
\itemize{
\item \code{plot_web_2D}: The function to plot all the results of the function \code{spiderweb()}
}}

\examples{
NULL
NULL
}
