% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_hyperparameters}
\alias{read_hyperparameters}
\title{Function to read hyperparameters and their values from the file}
\usage{
read_hyperparameters(input)
}
\arguments{
\item{input}{File name to input}
}
\value{
Parameters and their values
}
\description{
Function to read hyperparameters and their values from the file
}
\examples{
MaxWiK::MaxWiK_templates(dir = tempdir()) # See the templates and vignettes for usage
}
