% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restrict_data}
\alias{restrict_data}
\title{Function to restrict data in the size to accelerate the calculations}
\usage{
restrict_data(par.sim, stat.sim, stat.obs, size = 300)
}
\arguments{
\item{par.sim}{Data frame of parameters}

\item{stat.sim}{Data frame of outputs of simulations}

\item{stat.obs}{Data frame of observation point}

\item{size}{Integer number of points to leave from original dataset}
}
\value{
\code{restrict_data()} returns the list of: \cr
par.sim - restricted parameters which are close to observation point \cr
stat.sim - restricted stat.sim which are close to observation point
}
\description{
\code{restrict_data()} is based on rejection ABC method to restrict original dataset
}
\examples{
NULL
}
