% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_iKernel.R
\name{sampler_MaxWiK}
\alias{sampler_MaxWiK}
\title{Function to generate parameters and simulate a model based on MaxWiK algorithm}
\usage{
sampler_MaxWiK(
  stat.obs,
  stat.sim,
  par.sim,
  model,
  arg0 = list(),
  size = 500,
  psi_t,
  epsilon,
  nmax = 100,
  include_top = FALSE,
  slowly = FALSE,
  rate = 0.2
)
}
\arguments{
\item{model}{is a function to get output of simulation during sampling}

\item{arg0}{is a list with arguments for a model function, so that arg0 is NOT changed during sampling}

\item{size}{Number of point to restrict original dataset}

\item{nmax}{is maximal number of iterations}

\item{include_top}{Logical to include top points from \code{spider_web()} function to simulate or do not}

\item{slowly}{Logical for two algorithms: slow and fast seekers in sampling}

\item{rate}{Rate value in the range \code{[0,1]} to define
the rate of changing in the original top of sampled points}
}
\value{
\code{sampler_MaxWiK()} returns the list: \cr
results - results of simulations; \cr
best - the best value of parameter; \cr
MSE_min - minimum of MSE; \cr
number_of_iterations - number of iterations; \cr
time - time of sampling in seconds.
}
\description{
Function to generate parameters and simulate a model based on MaxWiK algorithm
}
\examples{
NULL
}
