% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{simulation_example}
\alias{simulation_example}
\alias{plot_3d_similarities}
\title{Example of simulation for lazy start}
\usage{
simulation_example(
  verbose = TRUE,
  to_plot = TRUE,
  seed = NA,
  model = c("Gaussian", "linear")[2],
  d = 2,
  x0 = c(3, 4),
  probability = TRUE,
  n = 1000,
  r = range(0, 10),
  psi = 4,
  t = 8,
  restrict_points_number = 300
)

plot_3d_similarities(sim, r = range(0, 10), n = 12)
}
\arguments{
\item{verbose}{Logical type to show or do not show messages during execution}

\item{to_plot}{Logical type to plot or do not plot graphical results of a simulation}

\item{seed}{Numeric type to set seed for a simulation, if seed = NA (by default) then it will be skipped}

\item{model}{Name of the toy model, can be 'Gaussian' or 'linear' only}

\item{d}{Integer number of dimension of the model}

\item{x0}{Truth value of parameter corresponding to observation point}

\item{probability}{Logical to apply the probabilistic distribution to input data generation}

\item{n}{Integer number of points for each axes to get net for plotly 3D graph}

\item{r}{Range of parameters to plot, by default \code{r = range(0, 10)}}

\item{psi}{Integer number. Size of each Voronoi diagram or number of areas/points in the Voronoi diagrams}

\item{t}{Integer number of trees in the Isolation Forest}

\item{restrict_points_number}{Integer number of points which will be considered in the simulation, \cr
other points will be skipped by rejection ABC}

\item{sim}{Results of the function \code{simulation_example}}
}
\value{
List of results of simulation with default values for all the parameters

Plot of 3D to show the similarity of each point of the parameter space
}
\description{
Example of simulation for lazy start
}
\section{Functions}{
\itemize{
\item \code{plot_3d_similarities}: Function to plot whole space of parameters
related to similarity of the observation point:
}}

\examples{
NULL
# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
sim = simulation_example( verbose = FALSE , to_plot = FALSE )
}
