% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_iKernel.R, R/experiments.R, R/methods.R
\name{sampler_MaxWiK}
\alias{sampler_MaxWiK}
\alias{make_hypersurface}
\alias{Get_call}
\alias{Get_parameter}
\alias{Get_call_all_methods}
\alias{Get_hyperparameters}
\title{Function to call a method to get a parameter estimation and MSE}
\usage{
sampler_MaxWiK(
  stat.obs,
  stat.sim,
  par.sim,
  model,
  arg0 = list(),
  size = 500,
  psi_t,
  epsilon,
  nmax = 100,
  include_top = FALSE,
  slowly = FALSE,
  rate = 0.2,
  n_simulation_stop = NA,
  check_err = TRUE
)

make_hypersurface(stat.obs, stat.sim, par.sim, size)

Get_call(
  method_name,
  kernel_name = "",
  dimension,
  iteration,
  stat.obs,
  stat.sim,
  par.sim,
  par.truth,
  model_function = Gauss_function,
  model_par = list(d = 1, x0 = 3, r = range(0, 10), noise = 0, A = 1, sigma = 1),
  hyper
)

Get_parameter(
  method_name,
  kernel_name = "",
  stat.obs,
  stat.sim,
  par.sim,
  hyper
)

Get_call_all_methods(
  dimension,
  iterations,
  stat.obs,
  stat.sim,
  par.sim,
  par.truth,
  cores = 4,
  model_function = Gauss_function,
  model_par = list(d = 1, x0 = 3, r = range(0, 10), noise = 0, A = 1, sigma = 1),
  hyper
)

Get_hyperparameters(stat.obs, stat.sim, par.sim, par.truth)
}
\arguments{
\item{stat.obs}{Data frame of statistics of observation point}

\item{stat.sim}{Data frame of statistics of simulations}

\item{par.sim}{Data frame of parameters}

\item{model}{Function to get output of simulation during sampling}

\item{arg0}{List with arguments for a model function, so that arg0 is NOT changed during sampling}

\item{size}{Number of points in the simulation based on MaxWiK algorithm}

\item{epsilon}{Criterion to stop simulation when \code{MSE_current - MSE_previous < epsilon}}

\item{nmax}{Maximal number of iterations}

\item{include_top}{Logical to include top points (network) from \code{spider_web()} function to simulate or do not}

\item{slowly}{Logical for two algorithms: slow and fast seekers in sampling}

\item{rate}{Rate value in the range \code{[0,1]} to define
the rate of changing in the original top of sampled points for slow scheme (if slowly = TRUE)}

\item{n_simulation_stop}{Maximal number of simulations to stop sampling.
If \code{n_simulation_stop = NA} then there is no restriction (by default).}

\item{check_err}{Logical parameter to check epsilon or do not}

\item{method_name}{Name of a method}

\item{kernel_name}{Name of kernel function}

\item{dimension}{Dimension of the model}

\item{iteration}{Iteration number of trial with the same model and other parameters}

\item{par.truth}{Truth parameter value to check result of estimation}

\item{model_function}{Function that is used as function to calculate output for an estimated parameter}

\item{model_par}{List of parameters for a model function}

\item{hyper}{List of hyperparameters to use for a method}

\item{cores}{Number of cores for parallel calculation with iterations}
}
\value{
\code{sampler_MaxWiK()} returns the list: \cr
\itemize{
\item results: results of all the simulations; \cr
\item best: the best value of parameter; \cr
\item MSE_min: minimum of MSE; \cr
\item number_of_iterations: number of iterations; \cr
\item time: time of sampling in seconds, \cr
\item n_simulations: the total number of simulations.
}

\code{sampler_MaxWiK()} returns the list: \cr

\code{Get_call()} returns the list: \cr
method_name = method_name, \cr
\itemize{
\item kernel_name, \cr
\item model_name, \cr
\item stochastic_term, \cr
\item MSE, \cr
\item running_time, \cr
\item iteration.
}

\code{Get_parameter()} returns the parameter estimation

\code{Get_call_all_methods()} returns data.frame with MSE for all defined methods

\code{Get_hyperparameters()} returns the list: \cr
\itemize{
\item list of epsilon and sigma for each kernel, \cr
\item tolerance.
}
}
\description{
Function to make 'hypersurface' or data set reduced from the bigger one with
selected points around observation point but
the points are away from each other
}
\section{Functions}{
\itemize{
\item \code{sampler_MaxWiK()}: Function to generate parameters and simulate a model based on MaxWiK algorithm

\item \code{make_hypersurface()}: Function to make hypersurface

\item \code{Get_parameter()}: Function to get a parameter estimation using a method

\item \code{Get_call_all_methods()}: Function to call all the methods to get estimation of parameter and MSE

\item \code{Get_hyperparameters()}: Function to get hyperparameters for a certain set of data

}}
\examples{
NULL
NULL
NULL
NULL 
NULL
}
