% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{Get_call}
\alias{Get_call}
\title{Function to call a method to get MSE}
\usage{
Get_call(
  method_name,
  kernel_name = "",
  model_name,
  stochastic_term,
  iteration,
  stat.obs,
  stat.sim,
  par.sim,
  G = NULL,
  par.truth
)
}
\arguments{
\item{method_name}{Name of a method}

\item{kernel_name}{Name of kernel function}

\item{model_name}{Name of a model}

\item{stochastic_term}{A number (usually in the range \code{[0,1]}) of stochastic term in the dataset \code{stat.sim}}

\item{iteration}{Iteration number of trial with the same model and other parameters}

\item{args}{List of arguments for a function to get MSE}
}
\value{
\code{Get_call()} returns the list: \cr
method_name = method_name, \cr
\itemize{
\item kernel_name, \cr
\item model_name, \cr
\item stochastic_term, \cr
\item MSE, \cr
\item running_time, \cr
\item iteration.
}
}
\description{
Function to call a method to get MSE
}
\examples{
NULL
}
