% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_iKernel.R, R/experiments.R
\name{sampler_method}
\alias{sampler_method}
\alias{sampler_all_methods}
\alias{Get_call}
\alias{Get_call_all_methods}
\alias{experiment_samplers}
\title{Function to call a method to get MSE}
\usage{
sampler_method(
  stat.obs,
  stat.sim,
  par.sim,
  model,
  method_name,
  kernel_name,
  arg0 = list(),
  size = 500,
  nmax = 30,
  G,
  model_name,
  dimension,
  stochastic_term,
  par.truth
)

sampler_all_methods(
  model_name,
  dimension,
  stochastic_term,
  stat.obs,
  stat.sim,
  par.sim,
  G,
  par.truth,
  cores = 4,
  nmax
)

Get_call(
  method_name,
  kernel_name = "",
  model_name,
  dimension,
  stochastic_term,
  iteration,
  stat.obs,
  stat.sim,
  par.sim,
  G = NULL,
  par.truth
)

Get_call_all_methods(
  model_name,
  dimension,
  stochastic_term,
  iterations,
  stat.obs,
  stat.sim,
  par.sim,
  G,
  par.truth,
  cores = 4
)

experiment_samplers(
  file_name = "./output.txt",
  model_name = "Gaussian",
  dimension = 6,
  stochastic_term = 5,
  rng = c(0, 10),
  restrict_points_number = 500
)
}
\arguments{
\item{stat.obs}{Data frame of statistics of observation point}

\item{stat.sim}{Data frame of statistics of simulations}

\item{par.sim}{Data frame of parameters}

\item{model}{is a function to get output of simulation during sampling}

\item{method_name}{Name of a method}

\item{kernel_name}{Name of kernel function}

\item{arg0}{is a list with arguments for a model function, so that arg0 is NOT changed during sampling}

\item{size}{Number of point to restrict original dataset}

\item{nmax}{is maximal number of iterations}

\item{G}{Matrix of similarities for K2-ABC based on isolation kernel}

\item{model_name}{Name of a model}

\item{dimension}{Dimension of the model}

\item{stochastic_term}{A number (usually in the range \code{[0,1]}) of stochastic term in the dataset \code{stat.sim}}

\item{par.truth}{Truth parameter value to check result of estimation}

\item{cores}{Number of cores for parallel calculation with iterations}

\item{iteration}{Iteration number of trial with the same model and other parameters}

\item{file_name}{File to save results}

\item{rng}{Range of points for each dimension, by default \code{rng = range(0,10)}}

\item{restrict_points_number}{Number of points which will be used in each method and that are close to observation point}
}
\value{
\code{sampler_method()} returns the list: \cr
results - results of simulations; \cr
best - the best value of parameter; \cr
MSE_min - minimum of MSE; \cr
number_of_iterations - number of iterations; \cr
time - time of sampling in seconds.

\code{sampler_all_methods()} returns data.frame with MSE for all defined methods

\code{Get_call()} returns the list: \cr
method_name = method_name, \cr
\itemize{
\item kernel_name, \cr
\item model_name, \cr
\item stochastic_term, \cr
\item MSE, \cr
\item running_time, \cr
\item iteration.
}

\code{Get_call_all_methods()} returns data.frame with MSE for all defined methods

\code{experiment_samplers()} returns data frame with results of experiment for sampling using all the methods
}
\description{
Function to call a method to get MSE
}
\section{Functions}{
\itemize{
\item \code{sampler_method()}: Function to generate parameters and simulate a model based on MaxWiK algorithm

\item \code{sampler_all_methods()}: Function to call all the methods to get estimation of parameter and MSE

\item \code{Get_call_all_methods()}: Function to call all the methods to get estimation of parameter and MSE

\item \code{experiment_samplers()}: Make experiments with sampling for all the methods and one case of toy model and dimension

}}
\examples{
NULL
NULL 
NULL
NULL 
NULL
}
