% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{Get_call}
\alias{Get_call}
\alias{Get_call_all_methods}
\title{Function to call a method to get MSE}
\usage{
Get_call(
  method_name,
  kernel_name = "",
  dimension,
  iteration,
  stat.obs,
  stat.sim,
  par.sim,
  G = NULL,
  model_par = list(name = "Gaussian", noise = 0, sigma = 1, A = 1)
)

Get_call_all_methods(
  model_name,
  dimension,
  stochastic_term,
  iterations,
  stat.obs,
  stat.sim,
  par.sim,
  G,
  par.truth,
  cores = 4,
  sigma = 1
)
}
\arguments{
\item{method_name}{Name of a method}

\item{kernel_name}{Name of kernel function}

\item{dimension}{Dimension of the model}

\item{iteration}{Iteration number of trial with the same model and other parameters}

\item{stat.obs}{Data frame of statistics of observation point}

\item{stat.sim}{Data frame of statistics of simulations}

\item{par.sim}{Data frame of parameters}

\item{G}{Matrix of similarities for K2-ABC based on isolation kernel}

\item{model_par}{List of parameters for a model function}

\item{model_name}{Name of a model}

\item{stochastic_term}{A number (usually in the range \code{[0,1]}) of stochastic term in the dataset \code{stat.sim}}

\item{par.truth}{Truth parameter value to check result of estimation}

\item{cores}{Number of cores for parallel calculation with iterations}
}
\value{
\code{Get_call()} returns the list: \cr
method_name = method_name, \cr
\itemize{
\item kernel_name, \cr
\item model_name, \cr
\item stochastic_term, \cr
\item MSE, \cr
\item running_time, \cr
\item iteration.
}

\code{Get_call_all_methods()} returns data.frame with MSE for all defined methods
}
\description{
Function to call a method to get MSE
}
\section{Functions}{
\itemize{
\item \code{Get_call_all_methods()}: Function to call all the methods to get estimation of parameter and MSE

}}
\examples{
NULL
NULL 
}
