% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{get_par_best_from_simnet}
\alias{get_par_best_from_simnet}
\title{Function to extract all the best parameters estimation from results of the function
\code{simulation_example_many_psi_t()}}
\usage{
get_par_best_from_simnet(simnet)
}
\arguments{
\item{simnet}{results of the function \code{simulation_example_many_psi_t()}}
}
\value{
\code{get_par_best_from_simnet()} returns all the best parameters estimation
}
\description{
Function to extract all the best parameters estimation from results of the function
\code{simulation_example_many_psi_t()}
}
\examples{
NULL
}
