% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_iKernel.R
\name{Get_parameter}
\alias{Get_parameter}
\title{Function to call a method to get parameter estimation and MSE for each used method}
\usage{
Get_parameter(
  method_name,
  kernel_name = "",
  stat.obs,
  stat.sim,
  par.sim,
  G = NULL,
  par.truth
)
}
\arguments{
\item{method_name}{Name of a method}

\item{kernel_name}{Name of kernel function}

\item{stat.obs}{Data frame of statistics of observation point}

\item{stat.sim}{Data frame of statistics of simulations}

\item{par.sim}{Data frame of parameters}

\item{G}{Matrix of similarities for K2-ABC based on isolation kernel}

\item{par.truth}{Truth parameter value to check result of estimation}
}
\value{
\code{par.truth,()} returns the list: \cr
method_name = method_name, \cr
\itemize{
\item kernel_name, \cr
\item model_name, \cr
\item stochastic_term, \cr
\item MSE, \cr
\item running_time, \cr
\item iteration.
}
}
\description{
Function to call a method to get parameter estimation and MSE for each used method
}
\examples{
NULL
}
