% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxWiK-package.R
\docType{package}
\name{MaxWiK-package}
\alias{MaxWiK}
\alias{MaxWiK-package}
\title{MaxWiK: Machine Learning Method Based on Isolation Kernel Mean Embedding}
\description{
Incorporates Approximate Bayesian Computation to get a posterior distribution and to select a model optimal parameter for an observation point. Additionally, the meta-sampling heuristic algorithm is realized for parameter estimation, which requires no model runs and is dimension-independent. A sampling scheme is also presented that allows model runs and uses the meta-sampling for point generation. A predictor is realized as the meta-sampling for the model output. All the algorithms leverage a machine learning method utilizing the maxima weighted Isolation Kernel approach, or 'MaxWiK'. The method involves transforming raw data to a Hilbert space (mapping) and measuring the similarity between simulated points and the maxima weighted Isolation Kernel mapping corresponding to the observation point. Comprehensive details of the methodology can be found in the papers Iurii Nagornov (2024) \doi{10.1007/978-3-031-66431-1_16} and Iurii Nagornov (2023) \doi{10.1007/978-3-031-29168-5_18}.
}
\author{
\strong{Maintainer}: Yuri Nagornov \email{nagornov.yuri@gmail.com} (\href{https://orcid.org/0000-0002-7935-6776}{ORCID}) [copyright holder]

}
\keyword{internal}
