% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_iKernel.R
\name{norm_vec}
\alias{norm_vec}
\alias{norm_vec_sq}
\title{The norm function for vector}
\usage{
norm_vec(x)

norm_vec_sq(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The squared root of sum of squared elements of the vector x or Euclid length of the vector x

The squared Euclid norm or the sum of squared elements of the vector x
}
\description{
The norm function for vector
}
\section{Functions}{
\itemize{
\item \code{norm_vec_sq}: The squared norm or the sum of squared elements of the vector x
}}

\examples{
norm_vec(c(3,4)) # that returns 5
norm_vec(c(12,5)) # that returns 13

norm_vec_sq(c(3,4)) # that returns 25
norm_vec_sq(c(12,5)) # that returns 169
}
