% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_iKernel.R
\name{sampler_MaxWiK}
\alias{sampler_MaxWiK}
\title{Function to generate parameters and simulate a model based on MaxWiK algorithm}
\usage{
sampler_MaxWiK(
  stat.obs,
  stat.sim,
  par.sim,
  model,
  arg0 = list(),
  size = 500,
  psi_t,
  epsilon,
  nmax = 100,
  include_top = FALSE,
  slowly = FALSE,
  rate = 0.2
)
}
\arguments{
\item{model}{Function to get output of simulation during sampling}

\item{arg0}{List with arguments for a model function, so that arg0 is NOT changed during sampling}

\item{size}{Number of points in the simulation based on MaxWiK algorithm}

\item{nmax}{Maximal number of iterations}

\item{include_top}{Logical to include top points (network) from \code{spider_web()} function to simulate or do not}

\item{slowly}{Logical for two algorithms: slow and fast seekers in sampling}

\item{rate}{Rate value in the range \code{[0,1]} to define
the rate of changing in the original top of sampled points for slow scheme (if slowly = TRUE)}
}
\value{
\code{sampler_MaxWiK()} returns the list: \cr
results - results of all the simulations; \cr
best - the best value of parameter; \cr
MSE_min - minimum of MSE; \cr
number_of_iterations - number of iterations; \cr
time - time of sampling in seconds, \cr
n_simulations - the total number of simulations.
}
\description{
Function to generate parameters and simulate a model based on MaxWiK algorithm
}
\examples{
NULL
}
