% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R, R/lib_iKernel.R, R/plotly.R
\name{simulation_example}
\alias{simulation_example}
\alias{simulation_example_many_psi_t}
\alias{get_Spider_MAP}
\alias{plot_3d_similarities}
\alias{plot_3d_net_similarities}
\title{Example of simulation for lazy start}
\usage{
simulation_example(
  verbose = TRUE,
  to_plot = TRUE,
  seed = NA,
  model = c("Gaussian", "linear")[2],
  d = 2,
  x0 = c(3, 4),
  probability = TRUE,
  n = 1000,
  r = range(0, 10),
  psi = 8,
  t = 8,
  restrict_points_number = 300
)

simulation_example_many_psi_t(
  verbose = TRUE,
  to_plot = TRUE,
  seed = NA,
  model = c("Gaussian", "linear")[2],
  d = 2,
  x0 = c(3, 4),
  probability = TRUE,
  n = 1000,
  r = range(0, 10),
  psi_t = data.frame(psi = c(4, 4, 8, 8, 8, 8, 10, 10), t = c(8, 20, 6, 8, 16, 20, 6,
    12)),
  restrict_points_number = 300,
  cores = 4
)

get_Spider_MAP(
  stat.sim,
  par.sim,
  stat.obs,
  restrict_points_number = 300,
  cores = 4,
  n_best = 8
)

plot_3d_similarities(sim, r = range(0, 10), n = 12)

plot_3d_net_similarities(simnet, r = range(0, 10), n = 20)
}
\arguments{
\item{verbose}{Logical type to show or do not show messages during execution}

\item{to_plot}{Logical type to plot or do not plot graphical results of a simulation}

\item{seed}{Numeric type to set seed for a simulation, if seed = NA (by default) then it will be skipped}

\item{model}{Name of the toy model, can be 'Gaussian' or 'linear' only}

\item{d}{Integer number of dimension of the model}

\item{x0}{Truth value of parameter corresponding to observation point}

\item{probability}{Logical to apply the probabilistic distribution to input data generation}

\item{n}{Integer number of points for each axes to get net for plotly 3D graph}

\item{r}{Range of parameters to plot, by default \code{r = range(0, 10)}}

\item{psi}{Integer number. Size of each Voronoi diagram or number of areas/points in the Voronoi diagrams}

\item{t}{Integer number of trees in the Isolation Forest}

\item{restrict_points_number}{Maximal number of points in the data sets to get MAP}

\item{psi_t}{Data.frame with different set of psi and t hyperparameters}

\item{cores}{Number of cores for parallel calculation}

\item{stat.sim}{Summary statistics of the simulations (model output)}

\item{par.sim}{\itemize{
\item data frame of parameters of the model
}}

\item{stat.obs}{Summary statistics of the observation point}

\item{n_best}{Number of best psi_t pairs adjusted for MaxWiK algorithm}

\item{sim}{Results of the function \code{simulation_example}}

\item{simnet}{Results of the function \code{simulation_example_many_psi_t}}
}
\value{
List of results of simulation with default values for all the parameters

List of results of simulation with default values for all the parameters

Maximum A Posteriori of meta-sampling distribution of parameters

Plot of 3D to show the similarity of each point of the parameter space

Many 3D plots to show the similarity of each point of the parameter space
for each set of hyperparameters psi and t
}
\description{
Example of simulation for lazy start
}
\section{Functions}{
\itemize{
\item \code{simulation_example_many_psi_t()}: Example of simulation for lazy start and different psi / t hyperparameters

\item \code{get_Spider_MAP()}: Function to get MAP of SpiderWeb algorithm based on different psi / t hyperparameters

\item \code{plot_3d_similarities()}: Function to plot whole space of parameters
related to similarity of the observation point:

\item \code{plot_3d_net_similarities()}: Function to plot whole space of parameters
related to similarity of the observation point for each set of hyperparameters psi and t

}}
\examples{
NULL
# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
sim = simulation_example( verbose = FALSE , to_plot = FALSE )
NULL
# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
sim = simulation_example_many_psi_t( verbose = FALSE , to_plot = FALSE )
NULL
# it takes a time for a simulation and then it will demonstrates results, \cr
# so, please, wait for a while
sim = simulation_example_many_psi_t( verbose = FALSE , to_plot = FALSE )
NULL
NULL
}
